Move MonthPagerAdapter to it's own class file
